+++ src/Resources/contao/languages/de/tl_fernschach_iccf_ratings.php
@@ -3,19 +3,20 @@
 /**
  * Backend-Modul: Übersetzungen im Eingabeformular
  */
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['name_legend'] = 'Personendaten';
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['surname'] = array('Nachname', 'Nachname des Spielers');
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['prename'] = array('Vorname', 'Vorname des Spielers');
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['country'] = array('Land', 'Land des Spielers');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['listId'] = array('Wertungsliste', 'Wertungsliste');
 
 $GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['iccf_legend'] = 'ICCF-Daten';
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['iccfid'] = array('ICCF-ID', 'ICCF-ID des Spielers');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['ratingDate'] = array('Wertungszahldatum', 'Beginndatum der Wertungsliste');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['title'] = array('Titel', 'ICCF-Titel des Spielers');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['rating'] = array('Wertungszahl', 'ICCF-Wertungszahl des Spielers');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['games'] = array('Partien', 'Anzahl der gewerteten Partien des Spielers');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['ratingDeviation'] = array('Wertungsabweichung', 'Wertungsabweichung (RD) des Spielers');
 
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['info_legend'] = 'Bemerkungen';
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['intern'] = array('Interne Hinweise', 'Interne Hinweise. Die Importfunktion schreibt hier Änderungen an den Stammdaten hinein.');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['flag_legend'] = 'Status';
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['flag'] = array('Kennzeichen', 'Kennzeichen (Flag): i = inaktiv');
 
 $GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['publish_legend'] = 'Veröffentlichung';
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['published'] = array('Veröffentlicht', 'Spieler aktivieren oder deaktivieren');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['published'] = array('Veröffentlicht', 'Wertungszahl aktivieren oder deaktivieren');
 
 /**
  * Buttons für globale Operationen
@@ -27,10 +28,9 @@ $GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['listen'] = array('Wertungslis
  * Buttons für Operationen
  */
 
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['new'] = array('Neuer Spieler', 'Neuen Spieler anlegen');
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['edit'] = array('Wertungszahlen des Spielers bearbeiten', 'Wertungszahlen des Spielers %s bearbeiten');
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['editheader'] = array('Spieler bearbeiten', 'Spieler %s bearbeiten');
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['copy'] = array('Spieler kopieren', 'Spieler %s kopieren');
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['delete'] = array('Spieler löschen', 'Spieler %s löschen');
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['toggle'] = array('Spieler aktivieren/deaktivieren', 'Spieler %s aktivieren/deaktivieren');
-$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['show'] = array('Spielerdetails anzeigen', 'Details des Spielers %s anzeigen');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['new'] = array('Neue Wertungszahl', 'Neue Wertungszahl anlegen');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['edit'] = array('Wertungszahl bearbeiten', 'Wertungszahl %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['copy'] = array('Wertungszahl kopieren', 'Wertungszahl %s kopieren');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['delete'] = array('Wertungszahl löschen', 'Wertungszahl %s löschen');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['toggle'] = array('Wertungszahl aktivieren/deaktivieren', 'Wertungszahl %s aktivieren/deaktivieren');
+$GLOBALS['TL_LANG']['tl_fernschach_iccf_ratings']['show'] = array('Wertungszahldetails anzeigen', 'Details der Wertungszahl %s anzeigen');